+++ plugins/prometheus/pkg/data/types.go
@@ -4,7 +4,7 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//     http://www.apache.org/licenses/LICENSE-2.0
+//	http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -41,10 +41,13 @@ type Statistics struct {
 // CombinedRequest is the cpu/memory and requests for a given pod/container
 type CombinedRequest struct {
 	Owner
-	cpu           []model.SamplePair
-	memory        []model.SamplePair
-	memoryRequest model.SampleValue
-	cpuRequest    model.SampleValue
-	memoryLimit   model.SampleValue
-	cpuLimit      model.SampleValue
+	cpu             []model.SamplePair
+	memory          []model.SamplePair
+	memoryRequest   model.SampleValue
+	cpuRequest      model.SampleValue
+	memoryLimit     model.SampleValue
+	cpuLimit        model.SampleValue
+	networkTransmit []model.SamplePair
+	networkReceive  []model.SamplePair
+	storageCapacity []model.SamplePair
 }