+++ plugins/rbac-reporter/CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+## 1.3.8
+* updated rbac schema
 ## 1.3.7
 * update alpine and x/net