+++ plugins/trivy/CHANGELOG.md
@@ -1,4 +1,6 @@
 # Changelog
+## 0.24.10
+* Allow insecure TLS for trivy using TRIVY_INSECURE env var
 
 ## 0.24.10
 * Fix for rolling scans when there are a lot of errors