+++ plugins/workloads/CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## 2.4.6
+* updated workload schema
+
 ## 2.4.5
 * update alpine and x/net